Add files via upload
This project aims to develop content in the COVID 19 category and 
also the Respiratory and Mental Health Sections of Physiopedia as a 
response to the COVID-19 pandemic. We intend to populate the site 
with practical, credible and -provoking information on all aspects of 
management of individuals with a diagnosis of COVID-19. The project 
was launched on when the World Health Organisation declared the 
 Pandemic. The first part of the project focused on 
creating content for the Coronavirus like active cases,recovery 
cases,death cases,total cases,prevention, WHO Guidelines and Vaccine 
Distribution Model.